+++ nova2d/space-shooter/src/enemies/Pawn1.cpp
@@ -15,7 +15,7 @@ namespace spaceshooter
 
 	void Pawn1::DisplayHit(int damage)
 	{
-		HitDisplay* hitDisplay = new HitDisplay("+ " + std::to_string(damage), "font4", "yellow", Rect(GetX() - GetWidth() / 2, GetY() - 16, 24, 16),
+		HitDisplay* hitDisplay = new HitDisplay("+ " + std::to_string(damage), "font4", "yellow", Rect(GetX() - (float)GetWidth() / 2.f, GetY() - 16.f, 24.f, 16.f),
 			Vec2(GetX(), GetY() - 128), 4000, 0);
 	}
